pyproject.toml
[project]
name = "gongyil-solver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11, <3.13"

dependencies = [
  # Use torch 2.4.0 and compatible vision/audio libraries
  "torch==2.4.0",
  "torchvision==0.19.0",
  "torchaudio==2.4.0",

  # Latest compatible versions of other core libraries
  "transformers>=4.41.0",
  "peft>=0.10.0",
  "accelerate>=0.30.0",
  "datasets>=2.19.0",
  "trl>=0.8.0",
  
  # Utilities
  "numpy>=1.26,<2",
  "scikit-learn>=1.0.0",
  "pandas>=1.3.0",
  "sentencepiece>=0.1.99",
  "mlx_lm==0.26.2",
  "google-generativeai>=0.8.5",
]
